refactor: split the class-object property arms out of get_field_by_name_tail
The `.name` arm pushed get_field_by_name_tail.rs past the 2000-line lint
cap. Move it and the adjacent #4949 `.prototype` arm into a
`class_object_props` sibling module; behavior unchanged (gap tests
re-verified byte-identical to node, perry-runtime 1321/0).
