+++ crates/perry-runtime/src/object/field_get_set/get_field_by_name_tail.rs
@@ -1312,40 +1312,15 @@ pub(crate) fn get_field_by_name_object_tail(
             let key_ptr = (key as *const u8).add(std::mem::size_of::<crate::StringHeader>());
             let key_len = (*key).byte_len as usize;
             let key_bytes = std::slice::from_raw_parts(key_ptr, key_len);
-            // #4949: heap class-expression values (`ClassExprFresh`) are real
-            // OBJECT_TYPE_CLASS objects, not INT32 class refs. Their `.prototype`
-            // read must still expose the live declared-class prototype object so
-            // tsc/tslib decorator code can inspect and mutate method descriptors.
-            if key_bytes == b"prototype"
-                && (*obj).object_type == crate::error::OBJECT_TYPE_CLASS
-                && (*obj).class_id != 0
-            {
-                let class_id = (*obj).class_id;
-                let value = super::super::class_registry::class_decl_prototype_value(class_id);
-                if value.to_bits() == crate::value::TAG_UNDEFINED {
-                    let value = super::super::class_prototype_ref_value(class_id);
-                    return JSValue::from_bits(value.to_bits());
+            // #4949 `.prototype` / #6497 `.name` on heap class-expression
+            // values — see `class_object_props`.
+            if (*obj).object_type == crate::error::OBJECT_TYPE_CLASS && (*obj).class_id != 0 {
+                if key_bytes == b"prototype" {
+                    return super::class_object_props::class_object_prototype_value(obj);
                 }
-                return JSValue::from_bits(value.to_bits());
-            }
-            // #6497: `.name` on a heap class-expression value (ClassExprFresh —
-            // #6470 also routes capturing function-body class DECLARATIONS
-            // through it) must expose the template's class name, same as the
-            // INT32 class-ref path's #2059 arm. An explicit `static name`
-            // member (an own field on the class object) still wins.
-            if key_bytes == b"name"
-                && (*obj).object_type == crate::error::OBJECT_TYPE_CLASS
-                && (*obj).class_id != 0
-            {
-                if let Some(v) = own_data_field_by_name(obj, key) {
-                    return v;
-                }
-                let class_id = (*obj).class_id;
-                if !super::super::class_registry::class_is_key_deleted(class_id, "name") {
-                    if let Some(cname) = super::super::class_registry::class_name_for_id(class_id) {
-                        let s =
-                            crate::string::js_string_from_bytes(cname.as_ptr(), cname.len() as u32);
-                        return JSValue::from_bits(crate::js_nanbox_string(s as i64).to_bits());
+                if key_bytes == b"name" {
+                    if let Some(v) = super::class_object_props::class_object_name_value(obj, key) {
+                        return v;
                     }
                 }
             }
